+++ src/lib/components/index.ts
@@ -89,8 +89,7 @@ import { LinkProps } from "./link/types";
 import { VuiList } from "./list/List";
 import { LOADING_BAR_COLOR, VuiLoadingBar } from "./loadingBar/LoadingBar";
 import { VuiMarkdown } from "./markdown/Markdown";
-import { VuiMarkdownEditor } from "./markdownEditor/MarkdownEditor";
-import { VuiMarkdownPreviewModal } from "./markdownEditor/MarkdownPreviewModal";
+import { VuiMarkdownPreviewModal } from "./markdownPreviewModal/MarkdownPreviewModal";
 import { VuiMenu } from "./menu/Menu";
 import { VuiMenuItem, MenuItem } from "./menu/MenuItem";
 import { VuiMenuList } from "./menuList/VuiMenuList";
@@ -298,7 +297,6 @@ export {
   VuiList,
   VuiLoadingBar,
   VuiMarkdown,
-  VuiMarkdownEditor,
   VuiMarkdownPreviewModal,
   VuiMenu,
   VuiMenuItem,
