--- src/lib/components/markdownEditor/_index.scss
@@ -1,96 +1,0 @@
-$markdownEditorPanelPadding: $sizeS;
-$markdownEditorLineHeight: 1.6;
-$markdownEditorRowHeight: $fontSizeSmall * $markdownEditorLineHeight;
-$markdownEditorMinRows: 2;
-$markdownEditorMaxRows: 24;
-
-.vuiMarkdownEditor {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  background-color: var(--vui-color-empty-shade);
-  border: 1px solid var(--vui-color-medium-shade);
-  border-radius: $sizeXxs;
-
-  &:focus-within {
-    outline: var(--vui-color-primary-shade) auto $sizeXxxs;
-    outline-offset: $sizeXxxs;
-  }
-}
-
-.vuiMarkdownEditor-isInvalid {
-  border-color: var(--vui-color-danger-shade);
-}
-
-.vuiMarkdownEditor-isDisabled {
-  background-color: var(--vui-color-light-shade);
-}
-
-.vuiMarkdownEditor__toolbar {
-  flex-shrink: 0;
-  padding: $sizeXs $sizeXxs $sizeXs $sizeXs;
-  border-bottom: 1px solid var(--vui-color-border-light);
-}
-
-// Owns the height for both tabs, so switching modes never resizes the editor. The height
-// itself is set inline from the rows prop; these bounds cap that value and the resize handle
-// alike, and past them the panels scroll rather than grow.
-.vuiMarkdownEditor__panels {
-  min-height: calc(#{$markdownEditorRowHeight} * #{$markdownEditorMinRows} + #{$markdownEditorPanelPadding * 2});
-  max-height: calc(#{$markdownEditorRowHeight} * #{$markdownEditorMaxRows} + #{$markdownEditorPanelPadding * 2});
-  resize: vertical;
-  // Required for resize to take effect, and clips each panel to the height set here.
-  overflow: hidden;
-
-  > [hidden] {
-    display: none;
-  }
-
-  > * {
-    height: 100%;
-  }
-}
-
-.vuiMarkdownEditor__textArea {
-  display: block;
-  height: 100%;
-  // The panel already enforces the height, so the textarea's own floor must not fight it.
-  min-height: 0;
-  border: none;
-  border-radius: 0;
-  background-color: transparent;
-  padding: $markdownEditorPanelPadding;
-  font-family: var(--vui-font-family-monospace);
-  font-size: $fontSizeSmall;
-  line-height: $markdownEditorLineHeight;
-
-  &:focus-visible {
-    outline: none;
-  }
-}
-
-.vuiMarkdownEditor__preview {
-  overflow-y: auto;
-  overscroll-behavior: contain;
-  scrollbar-width: thin;
-  padding: $markdownEditorPanelPadding;
-}
-
-.vuiMarkdownEditor__preview-isEmpty {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.vuiMarkdownPreviewModal__wordCount {
-  margin-left: $sizeXs;
-  font-size: $fontSizeSmall;
-  font-weight: $fontWeightNormal;
-  color: var(--vui-color-subdued-shade);
-}
-
-.vuiMarkdownPreviewModal__body {
-  max-width: 700px;
-  margin: 0 auto;
-  padding-top: $sizeS;
-}
