+++ src/lib/components/markdownPreviewModal/countWords.ts
