+++ src/lib/styles/styles.scss
@@ -35,7 +35,7 @@
 @import "../components/link/index";
 @import "../components/list/index";
 @import "../components/loadingBar/index";
-@import "../components/markdownEditor/index";
+@import "../components/markdownPreviewModal/index";
 @import "../components/menu/index";
 @import "../components/menuList/index";
 @import "../components/modal/index";
